Test vector cleanups
The changes in [1] harmonize `longfellow-zk`'s `zk_test.cc` with other
tests in that project. In particular, it addresses discrepancies in
Fiat-Shamir transcript initialization and pad generation. Additionally,
it augments the test so that it emits a test vector as JSON.

This corresponding change adopts the new test vector value thus
generated, and makes improvements to our own test vector code.

- Introduce a `test_vector` module refactored out of `circuit::tests`
- Remove hacks in `sumcheck::Prover`
- Add a yet-unused Ligero proof to test vector

[1]: https://github.com/tgeoghegan/longfellow-zk/tree/zk-test-cleanup
