+++ src/transcript.rs
@@ -74,7 +74,7 @@ impl Transcript {
     /// match longfellow-zk.
     pub fn initialize<FE: CodecFieldElement>(
         &mut self,
-        ligero_commitment: Option<&[u8]>,
+        ligero_commitment: &[u8],
         circuit: &Circuit,
         public_inputs: &[FE],
     ) -> Result<(), anyhow::Error> {
@@ -82,11 +82,8 @@ impl Transcript {
         // match longfellow-zk.
         // https://datatracker.ietf.org/doc/html/draft-google-cfrg-libzk-01#section-3.1.3
         // 3.1.3 item 1 says to append the "Prover message", "which is usually a commitment". In
-        // particular, a Ligero commitment. zk_test.cc does not do this, but zk_prover.h does when
-        // it constructs constraints, so we make it optional.
-        if let Some(ligero_commitment) = ligero_commitment {
-            self.write_byte_array(ligero_commitment)?;
-        }
+        // particular, a Ligero commitment.
+        self.write_byte_array(ligero_commitment)?;
 
         // 3.1.3 item 2: write circuit ID
         self.write_byte_array(&circuit.id)?;
